+++ eisbuk-admin/src/components/auth/PrivateRoute.tsx
@@ -2,10 +2,11 @@ import React from "react";
 import { Route, Redirect } from "react-router-dom";
 import { useSelector } from "react-redux";
 import { isLoaded, isEmpty } from "react-redux-firebase";
+
 import Unauthorized from "./Unauthorized";
 import Loading from "./Loading";
 
-const PrivateRoute = (props) => {
+const PrivateRoute: React.FC = (props) => {
   const auth = useSelector((state) => state.firebase.auth);
   const authInfoEisbuk = useSelector((state) => state.authInfoEisbuk);
   const amIAdmin =
